Stop populating SolverChoice due not trigger import time side-effects.
