+++ tests/plugins/conftest.py
@@ -13,7 +13,7 @@ from conda import plugins
 def plugin_manager(mocker):
     pm = pluggy.PluginManager('conda')
     pm.add_hookspecs(plugins)
-    mocker.patch('conda.base.context.get_plugin_manager', return_value=pm)
+    mocker.patch('conda.plugins.manager.get_plugin_manager', return_value=pm)
     return pm
